url_summarizer/requirements.txt
fastapi>=0.104.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
youtube-transcript-api>=0.6.1
beautifulsoup4>=4.12.2
python-multipart>=0.0.6
nltk>=3.8.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
anyio>=3.7.1